+++ Assets/Scripts/Factions/FactionCatalog.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using UnityEngine;
+
+[CreateAssetMenu(fileName = "FactionCatalog", menuName = "Clash of Pantheons/Faction Catalog")]
+public sealed class FactionCatalog : ScriptableObject
+{
+    [SerializeField] private FactionData[] factions;
+
+    public IReadOnlyList<FactionData> Factions => factions;
+}